refactor(codereview): simplify judge calibration and CI aggregation
- Move judge calibration cases to dataset/judge_calibration.jsonl with a loader, shrinking the Python module

- Reuse metrics.precision_recall in score_calibration instead of inline math

- Move JudgeCalibrationReport to types.py

- Drop the backward-compat single-run fallback in the pooled per-task F1 bootstrap CI
